Update qualityAnalysis.py
Various improvements to docstrings for various routines.

checkOverlaps(): added capability to explicitly list the lines to check instead of always checking all lines; added option to choose whether or not to plot map; added subtitle to plot to serve as legend; improved reporting of results.

checkXYPlan(): tidied code to make consistent with checkVert(); improved clarity and readability of report output.

checkVertPlan(): improved clarity and readability of report output.

checkSpeeds() / _reportSpeeds(): replaced clumsy and hard-to-follow algorithm with one very similar to that used in checkXYPlan().