Added interpolation to grid capability
Data in a whizz_file can now be interpolated to a regular grid by the following functions which rely on the capabilities of GMT in PyGMT:

grid_n_image(): Takes a list of channels, grids and images them.

whizz_to_array(): Writes vectors of x, y, and z to an xArrayDataset.

xarray_to_grid(): Takes an xArray Dataset and returns a grid of the z data.

image_pygmt(): Uses pygmt.grdimage() to image a grid.