Documentation, line selection options, more plots, bug-fix in checkVertPlan()
Many additions/improvements to docstrings.
Some improvements to reporting of results (including line number etc.)
Fixed bug in checkVertPlan() where coordinates of exceedance were incorrect for some line directions. Improved plotting outputs in same function.
Improved reporting in _reportSpeeds().
Made improvements to psdLineChannels().
Added specificLinesMap() principally to map all segments of a line but can be used to plot any list of lines to a map. Useful when there are lots of lines in a survey.
Added plotVertPlan() but this is not properly implemented yet.