minor improvements (detailed)
oddevenlines() - bringing grid attributes from the even grid across to the difference grid so that we have units and names for plot labels. Very simplistic but will do for now.

checkRepeatLines() - fixed a bug where different segments of a control line were treated as possible cross-overs.

added controls_lessthan_1000() which marks lines as controls if their value is < 1000. Works for some survey but not most!

checkIntersetionZ() - allowed input value of controls to default to all control lines. Only works if control line numbers are < 1000.