added auto-calculate algorithm to updateLineSpacing
updateLineSpacing can now use the position data of the flight-lines, controlled by the LineVariety attribute, to calculate the traverse and control line spacings.

A new function updateLineVariety has been added to allow manual setting of the LineVariety which previously could only be set by updateLineAttributes (using the line type). This allows a user to set the LineVariety update (to Control or Traverse) when the line type is not known, or not available.